Mirroring of workflows and calculations works
Either in groups, or not associated with any group.
Either sorted by groups, or in a top-level flat hierarchy.
"De-duplication" works by symlinking calculations if they are part
of a workflow.

Next, check what happens if a workflow is part of two groups -> Here,
de-deplucation should actually make more sense.